Add Unit Tests. WIP

+++ app/Http/Controllers/Api/v1/ApiController.php
@@ -3,12 +3,15 @@
 namespace App\Http\Controllers\Api\v1;
 
 use Illuminate\Http\Request;
+use Response;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\RecordLock;
+use App\User;
 use Input;
 use Auth;
 use DB;
+
 
 class ApiController extends Controller
 {
@@ -62,7 +65,7 @@ class ApiController extends Controller
   {
     //$this->middleware("checkaccess:{$this->model_short}.read");
     $this->middleware("checkaccess:{$this->model_short}.create",['only' => ['store'] ]);
-    $this->middleware("checkaccess:{$this->model_short}.update",['only' => ['show','update'] ]);
+    $this->middleware("checkaccess:{$this->model_short}.update",['only' => ['show','update','checkout'] ]);
     $this->middleware("checkaccess:{$this->model_short}.delete",['only' => ['destroy','destroyMany'] ]);
   }
 
@@ -95,11 +98,15 @@ class ApiController extends Controller
 
       array_push($this->with,'RevisionHistory');
 
-      $data = $model_class::with( $this->with)->findOrFail($id);
+      $model = $model_class::with( $this->with )->find($id);
 
-      $return = $data->toArray();
+      if ( ! $model ) {
+        return $this->notFound();
+      }
 
-      $return['readable_history'] = $this->getHistory($data);
+      $return = $model->toArray();
+
+      $return['readable_history'] = $this->getHistory($model);
 
       return response()->json( $return );
   }
@@ -139,7 +146,7 @@ class ApiController extends Controller
       }
     }
 
-    return $this->operationSuccessful();
+    return $this->operationSuccessful(201);
   }
 
   /**
@@ -156,6 +163,14 @@ class ApiController extends Controller
 
     $model_class = $this->model_class;
     $model = $model_class::find($ids);
+
+    if ( ! $model ) {
+      return $this->notFound();
+    }
+
+    if ( ! $model->isCheckedOutToMe() ) {
+      return $this->operationRequiresCheckout( $model );
+    }
 
     $model->update($input);
 
@@ -184,7 +199,7 @@ class ApiController extends Controller
       }
     }
 
-    return $this->operationSuccessful();
+    return $this->operationSuccessful(200);
 
   }
 
@@ -197,8 +212,18 @@ class ApiController extends Controller
   public function destroy($id)
   {
     $model_class = $this->model_class;
-    $model_class::find($id)->delete();
-    return $this->operationSuccessful();
+    $model = $model_class::find($id);
+
+    if ( ! $model ) {
+      return $this->notFound();
+    }
+
+    if ( $model->isCheckedOutToSomeoneElse() ) {
+      return $this->operationFailed("Delete Failed. That model is checked out to someone else.",409);
+    }
+
+    $model->delete();
+    return $this->operationSuccessful("{$this->model_short} record deleted.", 200);
   }
 
   /**
@@ -210,8 +235,19 @@ class ApiController extends Controller
   public function destroyMany()
   {
     $model_class = $this->model_class;
-    $model_class::whereIn('id', $this->getInputIds() )->delete();
-    return $this->operationSuccessful();
+    $models = $model_class::whereIn('id', $this->getInputIds() );
+
+    if ( ! $models ) {
+      return $this->notFound();
+    }
+
+    $count = $models->count();
+    $message = "{$this->model_short} record(s) deleted.";
+
+    $models->delete();
+
+
+    return $this->operationSuccessful(compact('message','count'), 200);
   }
 
   /**
@@ -301,8 +337,7 @@ class ApiController extends Controller
    */
   public function getInputIds()
   {
-    $input = Input::all();
-    return $input['ids'];
+    return Input::get('ids', null);
   }
 
   /**
@@ -311,8 +346,19 @@ class ApiController extends Controller
    * @param  int  $id
    * @return Response
    */
-  public function massUpdate($ids, $changes)
+  public function massUpdate($ids = null, $changes = null)
   {
+    $ids = $ids ?: $this->getInputIds();
+    $changes = $changes ?: Input::get('changes', null);
+
+    if ( ! $ids ) {
+      return $this->operationFailed('Mass update failed. Please specify what records to change and try again.', 406);
+    }
+
+    if ( ! $changes ) {
+      return $this->operationFailed('Mass update failed. Please specify what changes to make and try again.', 406);
+    }
+
     $table_name = (new $this->model_class)->getTable();
     DB::table($table_name)->whereIn('id',$ids)->update($changes);
     return $this->operationSuccessful();
@@ -420,27 +466,29 @@ class ApiController extends Controller
    */
   public function checkout($model,$id)
   {
-    $class = "\\App\\{$model}";
-    $item = $class::find($id);
-    $lock = $item->recordLock;
-
-    if( !Auth::user()->checkAccess( "{$model}.update" ) ) {
-      return $this->operationFailed( "You must be logged in as a user with {$model}.update permissions to do that." );
+    // make sure the user is allowed to checkout the record.
+    if( Auth::guest() || !Auth::user()->checkAccess( "{$model}.update" ) ) {
+      return $this->notAllowed( "Error checking out {$model} record. You must be logged in as a user with {$model} update permissions to do that." );
     }
 
-    if ( empty($lock) || $lock->checkExpired() || $lock->checkUser() ) {
-      RecordLock::create([
-        'lockable_id' => $item->id,
-        'lockable_type' => get_class($item),
-        'user_id' => \Auth::id()
-      ]);
+    $class = "\\App\\{$model}";
+    $item = $class::find($id);
 
-      return $this->operationSuccessful('Record Checked Out For Editing.');
+    if ( ! $item ) {
+      return $this->notFound();
+    }
+
+    $lock = $item->recordLock;
+
+    if ( empty($lock) || $lock->checkExpired() || $lock->checkUser() ) {
+      $item->checkoutToId( \Auth::id() );
+
+      return $this->operationSuccessful('Record Checked Out For Editing.', 200);
     }
 
     $user = User::find( $lock->user_id );
     $message = 'That record is checked out by : ' . $user->username;
-    return $this->operationFailed( $message );
+    return $this->operationFailed( $message, 410 );
   }
 
   /**
@@ -453,16 +501,21 @@ class ApiController extends Controller
   public function checkin($model,$id)
   {
     $class = "\\App\\{$model}";
-    $item = $class::findOrFail($id);
+    $item = $class::find($id);
+
+    if ( ! $item ) {
+      return $this->notFound();
+    }
+
     $lock = $item->recordLock;
 
     if ( !empty($lock) ) {
       $lock->delete();
-      return $this->operationSuccessful('Record Checked In.');
+      return $this->operationSuccessful('Record Checked In.', 200);
     }
 
     $message = 'That record is already checked in';
-    return $this->operationFailed( $message );
+    return $this->operationFailed( $message, 410 );
   }
 
   /**
@@ -470,16 +523,30 @@ class ApiController extends Controller
    * @method checkinAll
    * @return [type]     [description]
    */
-  public function checkinAll()
+  public function checkinAll($model = null)
   {
     $id = Auth::id();
+    $class = ( !! $model ) ? "App\\{$model}" : null;
 
-    if ( RecordLock::ofUser($id)->count() ) {
-      RecordLock::ofUser($id)->delete();
-      return $this->operationSuccessful('All Records Checked In.');
-    }  else {
-      return $this->operationFailed( 'Nothing to check in' );
+    //dd(RecordLock::all()->toArray());
+
+    if ( !! $class && RecordLock::ofUser($id)->ofType($class)->count() ) {
+      $message = "All {$model} records checked in.";
+      $count = RecordLock::ofUser($id)->ofType($class)->count();
+
+      RecordLock::ofUser($id)->ofType($model)->delete();
+      return $this->operationSuccessful( compact('message', 'count'), 200);
     }
+
+    if ( ! $class && RecordLock::ofUser($id)->count() ) {
+      $message = "All records checked in.";
+      $count = RecordLock::ofUser($id)->count();
+
+      RecordLock::ofUser($id)->delete();
+      return $this->operationSuccessful( compact('message', 'count'), 200);
+    }
+
+    return $this->operationFailed( 'Nothing to check in', 410 );
   }
 
   /**
@@ -487,12 +554,25 @@ class ApiController extends Controller
    * @method operationSuccessful
    * @return [type]              [description]
    */
-  public function operationSuccessful($message = false)
+  public function operationSuccessful($message = false, $statusCode = 200)
   {
-    return [
-      "errors" => false,
-      "message" => $message ?: "Operation Completed Successfully"
-    ];
+    if (is_numeric($message)) {
+      $statusCode = $message;
+      $message = null;
+    }
+
+    if(is_array($message)) {
+      $data = array_merge( [
+        "errors" => false
+      ], $message);
+    } else {
+      $data = [
+        "errors" => false,
+        "message" => $message ?: "Operation Completed Successfully"
+      ];
+    }
+
+    return Response::json($data,$statusCode);
   }
 
   /**
@@ -501,17 +581,65 @@ class ApiController extends Controller
    * @param  [type]          $e [description]
    * @return [type]             [description]
    */
-  public function operationFailed($e)
+  public function operationFailed($e, $statusCode = 404)
   {
+    if ( is_numeric($e) ) {
+        $statusCode = $e;
+        $e = null;
+    }
+
     if ( gettype($e) === 'string'  ) {
       $message = $e;
     } else {
       $message = "There was a problem completing your request :" . $e->getMessage();
     }
-    return [
+
+    $data = [
       "errors" => true,
       "message" => $message,
     ];
+
+    return Response::json($data, $statusCode);
+  }
+
+
+  /**
+   * Let the user know the operation they are attempting requires the model
+   *  be checked out first
+   * @method operationRequiresCheckout
+   * @param  [type]                    $model [description]
+   * @return [type]                           [description]
+   */
+  public function operationRequiresCheckout($model)
+  {
+    if ( $model->isCheckedOutToSomeoneElse() ) {
+      return $this->operationFailed( "Error performing action. Someone else has checked out that record.", 409 );
+    }
+    return $this->operationFailed("Error performing action. Checkout the record and try again.", 405);
+  }
+
+
+  /**
+   * Not found standard response
+   * @method notFound
+   * @return [type]   [description]
+   */
+  public function notFound()
+  {
+    return $this->operationFailed("Sorry, I was unable to find a {$this->model_short} with that id.", 404);
+  }
+
+
+  /**
+   * Not allowed standard response
+   * @method notAllowed
+   * @param  [type]     $permission [description]
+   * @return [type]                 [description]
+   */
+  public function notAllowed($message = null, $permission = "update")
+  {
+    $message = $message ?: "You must be logged in as a user with {$this->model_short} {$permission} permissions to do that.";
+    return $this->operationFailed( $message , 403 );
   }
 
 }
